+++ response/user.go
@@ -1,7 +1,7 @@
 package response
 
 type UserResponse struct {
-	ID               uint32 `json:"id"`
+	ID               uint64 `json:"id"`
 	Username         string `json:"username"`
 	Email            string `json:"email"`
 	IsEmailConfirmed bool   `json:"isEmailConfirmed"`
